fix(gateway): settlement-first close for readback-only recovery; determinize state-enrichment CI (PR #152 review CF1 residual)
Close now lets every transaction whose step was failed by close unwind
and settle ITS OWN terminal before the deadline sweep (bounded drain,
microtask-time). The sweep therefore only classifies transactions parked
in a user hook, where the dispatch-phase bit stays authoritative. A
readback-only recovery whose authoritative GET is interrupted before
dispatch classifies GatewayClosedError → one interrupted terminal with
a failed caller outcome; with the GET already running it settles
unknown/unknown — recovery never again reports interrupted while its
caller reports unknown.

Barrier matrix (tests/github-write-close.test.ts): queued lease, marker
hook, pacing, running dispatch, running readback and unknown-budget-
queued recovery each assert the caller semantics, exactly one terminal,
and the zero/maybe-write classification.

CI flake root cause (state-enrichment): a mid-parallel aggregate page
failure lets the scheduler legitimately drop the orphaned sibling page
(the logical request already terminaled), which made the dispatch
multiset microtask-order dependent. The fake now fails the aggregate
only AFTER both pages ran (non-list issues body), deterministically.

Late-response diagnostics moved to gateway-lifecycle.ts; close-barrier
tests split into their own suite (owner back under the 800-line hard
limit).
