+++ tests/github-gateway-writes.test.ts
@@ -629,154 +629,3 @@ test('review F1: the dependency-unlock marker scan finds an existing comment on 
   assert.equal(outcome.outcome, 'confirmed')
   assert.equal(posts, 0, 'a marker beyond page 1 must still suppress the duplicate POST')
 })
-
-test('review F2: close() settles a queued write with one interrupted terminal and never strands the lease', async () => {
-  const owner = createGithubGatewayOwner()
-  const ctx = {
-    shell: {
-      resolve: (spec: unknown) => spec,
-      run: async () => {
-        throw new Error('must not dispatch — the gateway is closed')
-      },
-    },
-  }
-  const reader = new GithubRestReader(ctx as never, { owner, minimumIntervalMs: 0 })
-  let releaseA: (() => void) | null = null
-  const gateA = new Promise<void>((resolve) => {
-    releaseA = resolve
-  })
-  const spec: GithubWriteSpec<{ repoKey: string }, unknown> = {
-    id: 'test-gated-write',
-    keys: () => ['o/r/issues/5', 'repo:o/r'],
-    priority: 'normal',
-    deadlineMs: 5_000,
-    maxPages: 2,
-    repeatable: true,
-    dispatch: async () => {
-      await gateA
-      return {}
-    },
-    readback: {
-      run: (read) => read.json('repos/o/r/issues/5'),
-      confirms: () => true,
-    },
-  }
-  const writeA = runWriteTransaction(owner, reader, spec, { repoKey: 'o/r' })
-  await new Promise((resolve) => setTimeout(resolve, 30))
-  const writeB = runWriteTransaction(owner, reader, spec, { repoKey: 'o/r' })
-  await new Promise((resolve) => setTimeout(resolve, 30))
-  const closing = owner.close({ drainMs: 100 })
-  await assert.rejects(() => writeB, /排队写请求被中断/, 'the queued write is interrupted, not hanging')
-  releaseA?.()
-  const outcomeA = await writeA
-  assert.equal(
-    outcomeA.outcome,
-    'unknown',
-    'the running write settles unknown once its readback hits the closed gateway',
-  )
-  await closing
-  const events = owner.lifecycleEvents()
-  const bDeclared = events.filter((event) => event.kind === 'declared')
-  assert.equal(bDeclared.length, 2, 'both writes are lifecycle-visible before they block')
-  const terminals = events.filter((event) => event.kind === 'terminal')
-  assert.equal(terminals.length, 2, 'exactly one terminal per logical request')
-  const outcomes = terminals.map((event) => (event as { outcome: string }).outcome).sort()
-  assert.deepEqual(
-    outcomes,
-    ['interrupted', 'unknown'],
-    'the queued write is interrupted; the dispatch-attempted write is swept unknown (it may have executed)',
-  )
-})
-
-test('review CF1: a transaction parked in its marker is swept interrupted and later fails with zero dispatch', async () => {
-  const owner = createGithubGatewayOwner()
-  const ctx = {
-    shell: {
-      resolve: (spec: unknown) => spec,
-      run: async () => {
-        throw new Error('must not dispatch — the gateway closed before the marker resolved')
-      },
-    },
-  }
-  const reader = new GithubRestReader(ctx as never, { owner, minimumIntervalMs: 0 })
-  let releaseMarker: (() => void) | null = null
-  const markerGate = new Promise<void>((resolve) => {
-    releaseMarker = resolve
-  })
-  const spec: GithubWriteSpec<{ repoKey: string }, unknown> = {
-    id: 'test-marker-parked',
-    keys: () => ['o/r/issues/5', 'repo:o/r'],
-    priority: 'normal',
-    deadlineMs: 5_000,
-    maxPages: 2,
-    repeatable: false,
-    dispatch: (read) => read.mutate('repos/o/r/issues/5/comments', 'POST', { body: 'x' }),
-    readback: {
-      run: (read) => read.json('repos/o/r/issues/5'),
-      confirms: () => true,
-    },
-  }
-  const transaction = runWriteTransaction(
-    owner,
-    reader,
-    spec,
-    { repoKey: 'o/r' },
-    {
-      persistMarker: async () => {
-        await markerGate
-      },
-    },
-  )
-  await new Promise((resolve) => setTimeout(resolve, 30))
-  await owner.close({ drainMs: 50 })
-  const terminals = owner.lifecycleEvents().filter((event) => event.kind === 'terminal')
-  assert.equal(terminals.length, 1, 'the swept terminal is the only answer')
-  assert.equal((terminals[0] as { outcome: string }).outcome, 'interrupted', 'pre-dispatch writes are interrupted')
-  releaseMarker?.()
-  const outcome = await transaction
-  assert.equal(outcome.outcome, 'failed', 'the caller sees the write provably never happened')
-  assert.equal(owner.lifecycleEvents().filter((event) => event.kind === 'terminal').length, 1)
-})
-
-test('review CF1: a transaction parked in its readback is swept unknown — it may have executed', async () => {
-  const owner = createGithubGatewayOwner()
-  let releaseReadback: (() => void) | null = null
-  const readbackGate = new Promise<void>((resolve) => {
-    releaseReadback = resolve
-  })
-  const ctx = {
-    shell: {
-      resolve: (spec: unknown) => spec,
-      run: async (spec: { command: string }) => {
-        if (spec.command.includes('--method POST'))
-          return { exitCode: 0, stdout: { text: 'HTTP/1.1 201\n\n{"id":1}' }, stderr: { text: '' } }
-        await readbackGate
-        return { exitCode: 0, stdout: { text: 'HTTP/1.1 200\n\n{}' }, stderr: { text: '' } }
-      },
-    },
-  }
-  const reader = new GithubRestReader(ctx as never, { owner, minimumIntervalMs: 0 })
-  const spec: GithubWriteSpec<{ repoKey: string }, unknown> = {
-    id: 'test-readback-parked',
-    keys: () => ['o/r/issues/5', 'repo:o/r'],
-    priority: 'normal',
-    deadlineMs: 5_000,
-    maxPages: 2,
-    repeatable: true,
-    dispatch: (read) => read.mutate('repos/o/r/issues/5/comments', 'POST', { body: 'x' }),
-    readback: {
-      run: (read) => read.json('repos/o/r/issues/5'),
-      confirms: () => true,
-    },
-  }
-  const transaction = runWriteTransaction(owner, reader, spec, { repoKey: 'o/r' })
-  await new Promise((resolve) => setTimeout(resolve, 30))
-  await owner.close({ drainMs: 50 })
-  const terminals = owner.lifecycleEvents().filter((event) => event.kind === 'terminal')
-  assert.equal(terminals.length, 1)
-  assert.equal((terminals[0] as { outcome: string }).outcome, 'unknown', 'a dispatched write is never guessed closed')
-  releaseReadback?.()
-  const outcome = await transaction
-  assert.equal(outcome.outcome, 'unknown', 'caller and lifecycle agree on the single unknown answer')
-  assert.equal(owner.lifecycleEvents().filter((event) => event.kind === 'terminal').length, 1)
-})
